Update README.md
The main purpose of this project is to build a face recognition-based attendance monitoring
system for educational institution to enhance and upgrade the current attendance system into
more efficient and effective as compared to before. The current old system has a lot of
ambiguity that caused inaccurate and inefficient of attendance taking. Many problems arise
when the authority is unable to enforce the regulation that exist in the old system. The
technology working behind will be the face recognition system. The human face is one of the
natural traits that can uniquely identify an individual. Therefore, it is used to trace identity as
the possibilities for a face to deviate or being duplicated is low. In this project, face databases
will be created to pump data into the recognizer algorithm. Then, during the attendance taking
session, faces will be compared against the database to seek for identity. When an individual
is identified, its attendance will be taken down automatically saving necessary information into
a log.txt file.